Fix font:raleway version mess
This font is a mess because the upstream sources are all over the place.
Each new version seems to show up on a different distribution channel,
usually without a version number or source code, and people have been
just guessing what they are. There are several upstream source
repositories posted by different people that have worked on the family
but they don't even all match each other.

The most coherent clues come from the changelog posted at Google Fonts:

https://github.com/google/fonts/blob/master/ofl/raleway/FONTLOG.txt

There it is clear that some version scheme has been attempted. The
personal sandbox of one of the designers also suggests the version
3.x being the last official release.

https://github.com/impallari/Raleway

There he also has some ongoing development work tagged as 4.x, but as of
yet it is incomplete and considered unreleased.

This tries to straighten the version scheme processing so snapshots are
ignored, devel versions are so tagged, and the 1.x, 2.x and 3.x releases
are actually assumed to be correct.

Note also Raleway Dots is a separate font with a different upstream
release cycle not actually versioned along with Raleway.