Combine various Gitlab packages that are release-synced
The Community Edition (CE) and Enterprise Edition (EE) are now code
synced and releases are always together. Grouping them as a meta package
makes sense to me because the versions should always be the same even if
the file sets and licenses vary.

Likewise the various runner related packages (some of which have names
from before an upstream rename) are synced between regular and docker
image editions, so lumping them together makes sense to me.