Ignore Alpine Linux git snapshots.
The recommended way to version git snapshots for projects that have no
releases yet is to use the version pattern: `0_git<commitdate>`. This
way, any actual released version will be considered newer than these
snapshots.

Ignore any version using this pattern as the date is not always
reliable (some use the current date), so it's not a good way to compare.