Sink 20YY for AdaCore packages
AdaCore seems to have switched version schemes in a way: They used
publish the GNATCOLL libraries and the gprbuild tool with a year as
the version number. Starting with 2021 they seem to use the last two
digits of the year followed by two version parts which have all been
zero so far. This can be seen from the GitHub release pages:

* https://github.com/AdaCore/gnatcoll-core/releases/
* https://github.com/AdaCore/gnatcoll-db/releases/
* https://github.com/AdaCore/gnatcoll-bindings/releases/
* https://github.com/AdaCore/gprbuild/releases

The docs seem to use an YY.0w scheme I can't quite figure out,
but this is unused by any package repository, e.g.

* https://docs.adacore.com/gnatcoll-docs/
* https://docs.adacore.com/live/wave/gnatcoll-core/html/gnatcoll-core_ug/index.html
* https://docs.adacore.com/live/wave/gnatcoll-iconv/html/gnatcoll-iconv_ug/index.html
* https://docs.adacore.com/live/wave/gnatcoll-syslog/html/gnatcoll-syslog_ug/index.html
* https://docs.adacore.com/live/wave/gnatcoll-readline/html/gnatcoll-readline_ug/index.html
* https://docs.adacore.com/gprbuild-docs/html/gprbuild_ug.html

The verlonger rule for gprbuild is removed because the currently
official version numbers have more than one component.