Merge flavors into gnatcoll-bindings, gnatcoll-db
The AdaCore repositories

* https://github.com/AdaCore/gnatcoll-bindings
* https://github.com/AdaCore/gnatcoll-db

contain a number of sublibraries (and executables) that are
developed and released together. Some repositories split these
up into their respective subprojects (nixpkgs, AUR) others
package them all together (Fedora). For the former the projects
gnatcoll-db and gnatcoll-bindings would not exist whereas it
would be the only packages existing in the latter repositories.

To alleviate this, we just merge them all into those two names
with added flavors. The flavor annotations are made manually
since the parent projects name is not the prefix of the
subprojects' names.

(gnatcoll-db2ada and gnatinspect are currently named incorrectly
in nixpkgs, but this will be fixed with
https://github.com/NixOS/nixpkgs/pull/201026/commits/ad3c52bf07a1c1295840fe3311cd53c390a6ab39)