sdl*: rework merge/split rules
Instead of hacky merge rules which try to distinguish sdl* versions,
switch to straightforward merge and split by version.

Note: the tricky one is sdl-ttf vs. sdl2-ttf, as these share major
version. But it seems that we can just split after 2.0.11 release which
is the latest for sdl-ttf. This even fixes debian, in sdl-ttf2.0 which
is classified sdl2-ttf by its name, is in fact sdl-ttf.